DOC : Navbar #6 reorganize components folder structure
