+++ src/routes/mainRoutes.tsx
@@ -1,6 +1,5 @@
 import { Navigate } from 'react-router-dom';
-import AuthLayout from '../layouts/AuthLayout';
-import AnonymousLayout from '../layouts/AnonymousLayout';
+import { AuthLayout, AnonymousLayout } from '../components';
 import {
     Login,
     Signup,
